backend/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
base58==2.1.1
bitarray==2.8.2
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.1
constantly==23.10.4
cryptography==41.0.5
cytoolz==0.12.2
daphne==4.0.0
Django==4.2.6
django-environ==0.11.2
djangorestframework==3.14.0
ecdsa==0.18.0
eth-abi==4.2.1
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.1
eth-utils==2.3.0
frozenlist==1.4.0
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.1
hexbytes==0.3.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
mysqlclient==2.2.0
parsimonious==0.9.0
protobuf==4.24.4
py-bip39-bindings==0.1.11
py-ed25519-zebra-bindings==1.0.1
py-sr25519-bindings==0.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
PyNaCl==1.5.0
pyOpenSSL==23.3.0
pytz==2023.3.post1
pyunormalize==15.0.0
redis==5.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.10.6
scalecodec==1.2.7
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
substrate-interface==1.7.4
toolz==0.12.0
Twisted==23.8.0
txaio==23.1.1
types-pkg-resources==0.1.3
typing_extensions==4.8.0
urllib3==2.0.7
web3==6.11.1
websocket-client==1.6.4
websockets==12.0
wsaccel==0.6.6
xxhash==3.4.1
yarl==1.9.2
zope.event==5.0
zope.interface==6.1
